[dynamic-shapes] Add support for concatenate
In order to add support for concatenating dynamic shapes I add
a new shape parameter to the concatenate primitive to store the
output_shape.

We pass the shape are a parameter so that it can be computed once,
while lowering lax.concatenate to primitives. At that time we are
in a context where we can stage computations with shapes. This allows
us to keep the staging and typechecking rules free of shape computations.