Don't capture all exceptions from platform initialization.
Add a new exception class PlatformInitializationError that is treated as a non-fatal failure and is caught. All other exceptions will be propagated to the user.

This change avoids the problem of bugs in platform initialization code being silently ignored because of an over-general `Exception` guard.

Also permit the creation of platforms that have zero devices, just don't permit them to be the default platform.

PiperOrigin-RevId: 466039205