[jax2tf] Narrows down the dependencies from jax2tf on TF.
This breaks a cyclic dependency when depending on JAX from TFLite.

Context: I am working on a JAX-->TFLite converter in the TFLite codebase, so I have to add a dependency TFLite --> JAX, but jax2tf itself imports TF, which in turns depends on TFLite again. So we have a cyclic dependency: TFLite --> JAX --> TF --> TFLite.

As discussed with the TF team, the real problem is the TF should not pull in TFLite, but fixing this may take a while, so this PR is a stopgap solution.

PiperOrigin-RevId: 469484696