Use reduce_window for faster cumulative reductions on small inputs on GPU.
On GPU, reduce_window is executed in a single fusion and associative_scan
is split into multiple to materialize intermediate calculations.
On small inputs reduce_window is faster being a single fusion,
but on larger ones is slower because of O(n^2) complexity.
The conservative value of the threshold to choose between the two algorithms was obtained by benchmarking.
cumred_tpu_impl was renamed into cumred_reduce_window_impl to reflect that it is useful not only on TPUs.

PiperOrigin-RevId: 472504553