[jax2tf] Updates to the documentation
One change I made throughout is to use the verb lower instead of
convert when refering to what jax2tf does. This is more accurate,
because what is happening is that the lowering to HLO is replaced
with a lowering to TF ops. Some users were confused by the use
of the verb convert as suggesting some sort of source-to-source
translation of JAX programs to TF, which sounds very complicated
and risky.