Minimally support `device` argument on `jit` in the `jax.Array` path
This means that only a single device is allowed to flow through this path. This is a compromise i.e. it will support the existing codepaths but won't support sharded arrays to go through this path and encourage users to use other well supported techniques like using device_put explicitly instead of relying on `jit` to do that for you.

PiperOrigin-RevId: 473373822