Use `d.process_index == d.client.process_index()` to find addressable devices from a list of addressable and non-addressable devices.
For GPU and TPU, this is a no-op change.

For CPU, this makes a difference in multiprocess computation because if you run a `jit` on a CPU when you are on process 1 (for example), the `process_index` on the CPU device is 0. This is wrong since the CPU device is local to process 1 but the fix in the CL is a workaround until a runtime fix lands.

PiperOrigin-RevId: 484403748