[jax2tf] Fix native lowering for modules that don't use some inputs.
We must drop the corresponding actual arguments when invoking the function
and we must also be careful to compute the dim_args_specs based only
on the kept inputs.

As a side benefit, we now allow dimension variables that occur in
more complex polynomials, as long as they also occur as trivial
monomials somewhere in the input shapes for the kept arguments.