Migrate JAX from producing MHLO to producing StableHLO
As discussed over the last few months, it is desirable to migrate JAX from producing MHLO to producing StableHLO, and this CL makes this happen. More specifically:
  1) MLIR lowerings now produce StableHLO ops instead of MHLO ops.
  2) Fallback lowerings now produce StableHLO ops as well.
  3) Occurrences of "MHLO" in prose have been changed to "StableHLO", unless the documents are immutable (changelog, JEPs).

From time to time, it might be useful to produce MHLO directly, so MHLO is not going away and is still within arm's reach (although compatibility guarantees will only be provided for StableHLO and not for MHLO):
  a) `from jax._src.lib.mlir.dialects import mhlo` still does the same thing.
  b) `XlaLowering.mhlo()` is available as well, but its implementation has changed - it calls `stablehlo-legalize-to-hlo` underneath.
  c) `Lowering.as_text()/compiler_ir()` still support `dialect="mhlo"`, but the default has changed to "stablehlo".
  d) We're still using `mhlo.is_same_data_across_replicas` and `mhlo.sharding` because StableHLO currently lacks comparable functionality. https://github.com/openxla/stablehlo/issues/744 tracks the corresponding work, but it is not a blocker - we can use these attributes with StableHLO without any issues.

Special care has been taken to make these changes in a backward-compatible fashion. Instead of calling `mhlo.FooOp` or `stablehlo.FooOp` directly, the code now calls `xhlo.FooOp` where `xhlo = stablehlo if xla_client.mlir_api_version >= 40 else mhlo`. I hope that this kind of new terminology looks good to you all.

There's one more thing that I'd like to bring up. I've run a code formatter on the CL to make sure that newly written code is formatted properly, and since this CL changes quite a bit, the formatter got its hands on a lot of code and made a bunch of changes. I don't know which of these changes are appropriate and which ones aren't - please provide guidance on what we should keep and what I should revert.

PiperOrigin-RevId: 494578385