Don't tuple high-arity computations on GPU.
I can think of no reason we do this, and it seems strictly preferable to not have to form an on-device tuple since it requires a H2D copy.

As best I can reconstruct it, the sequence of events that got us to the current state was:
a) we had backend-independent logic to tuple computations with > 100 arguments. This logic was motivated only by TPU.
b) we disabled tupling on CPU, because the XLA:CPU folks noticed it wasn't needed.
c) we increased the argument limit on TPU to 2000, because the TPU compiler folks noticed it was too low.

This left GPU with the old TPU behavior for no good reason.

PiperOrigin-RevId: 496993803