[jax2tf] Simplify _eval_shape for the constant dimensions
Before this change _eval_shape would return a tf.Const for the constant dimensions in the shape. Now it returns integer constants.

This change reduces the size of the graph for the common cases when the dimensions are constant. It should not be a necessary change though but on some TPU flavors this results in TF2XLA bridge return error "XLA can't use dynamic begin values for slice" even when the dimension is actually a constant.

PiperOrigin-RevId: 497986554