[call_tf] WIP: find a better mechanism for dealing with DCE
The CallTfEffect was added in cl/496022997 to work around DCE
removing arguments to call_tf. This may have been a convenient
available mechanism for a quick fix, but it is not right to
highjack the effects mechanism. In particular (b/242119183)
users of AOT APIs now can get the impression that a computation
that calls call_tf has externally visible effects.

PiperOrigin-RevId: 501220005