checkify: in traceable, convert constvars to args b/c they may be tracers
Because the 'consts' that come out when we trace the user function to a jaxpr
could be `Tracers` (e.g. doing `grad`-of-`checkify`), **by `ClosedJaxpr`
convention** we shouldn't package them up in a ClosedJaxpr.

Since checkify isn't a higher-order primitive, and is instead an initial-style
transformation, putting Tracers in the ClosedJaxpr **was not actually a bug**,
in that it wouldn't cause any incorrect behavior. Still, it's _usually_ a bug
for Tracers to end up in ClosedJaxprs, since _usually_ ClosedJaxprs are used
for higher-order primitives. For that reason, I'd like to add the assertion in
core.py.

Without this change, if we add the assertion in core.py, the test
AssertPrimitiveTests.test_assert_discharging_no_data_dependence fails.