Eagerly register backend factories for PJRT C API plugins.
The prior mechanism was triggering the xla_client (in jaxlib) to load the plugins based on the environment variable, but there was nothing making them known to the JAX-side xla_bridge. Further, they were all being loaded regardless of what plugin was being requested.

This reworks the mechanism to eagerly register backend factories with the JAX-side xla_bridge so that they are availble when requested (i.e. as part of JAX_PLATFORMS env var or equivalent).

There is an opaque `TODO(b/237099479): Move to xla_bridge.py when ready.` in xla_client.py that can't be seen from the OSS side, so not sure what the intent here is. As near as I can tell, this never worked and is a temporary measure while we work out a better mechanism.