- Implement circular buffer - Circular buffer test
Since RoPE is a function s.t. f(x_m, m) * f(x_n, n) == f(x_m * x_n, m-n) it is therefore only a function of current length when we write it into the buffer, therefore the only require modifications are in the circular buffer itself.

Offset would be cleaner as an integer, but shard_map does not yet take 'None' as a sharding spec (at the least P(None)), and therefore requires rank 1 inputs. Will fix once this is addressed.

PiperOrigin-RevId: 507395004