[jax2tf] Customize limitations for native lowering mode
In the native lowering mode we expect no discrepancies between JAX native and jax2tf. We change the jax2tf limitations to be customizeable whether they are
applicable to native lowering or not. It turns out that almost all limitations are for non-native lowering mode.

An interesting case that arose was the min/max behavior under NaN. Even in native lowering, JAX and TF use different values for the compilation flag xla_cpu_enable_fast_min_max, which results in different results in presence of NaN.

PiperOrigin-RevId: 511078114