[jax2tf] Enable strict platform checking for native serialized modules.
This takes advantage of recent changes to XlaCallModule that allow
us to use a `platforms` attribute to specify what are the allowable
platforms for a serialized module.

Also add a `experimental_native_lowering_strict_checks` parameter
to `jax2tf.convert` to disable the check. The flag is False for
now (backwards compatible).