Refactor bcoo_dot_general GPU lowering
The goal of this is to make it easier to address the out-of-bound index issue. Our current GPU logic grew somewhat organically over time, and the logic for which sub-routine is called is spread over multiple locations. This change updates the branching such that the logic for each sub-routine appears directly adjacent to its call site; the tradeoff is that other considerations (such as whether to raise a warning) have to be duplicated between the cases.

Additionally, I simplified some of the hlo operation calls to make the code easier to follow.

PiperOrigin-RevId: 513025719