[sparse] simplify bcoo_dot_general GPU lowering rule
Also, remove the cusparse lowering for batched matmul, because in testing I found that it returns incorrect results for CUSPARSE_SPMM_COO_ALG4. Our tests haven't revealed that because we currently only test for a single batch. To re-land this, we can add it to the private primitives in _lowerings and add another elif clause in the GPU impl.

PiperOrigin-RevId: 513604587