in `closure_convert`, hoist all tracers involved in staging
We need to hoist all `DynamicJaxprTracer`s in `jax.closure_convert`,
not only those that are float-dtyped. Leaving any `DynamicJaxprTracer`
in the closure of a `custom_{jvp,vjp}`'ed function might result in a
tracer leak, as staging to jaxpr could already be complete by the time
the closure is consumed.

Co-authored-by: Matthew Johnson <mattjj@google.com>