[jax2tf] Rename experimental_native_lowering to native_serialization
We refer to the feature as serialization rather than just lowering,
because the former is both more widely understood and is actually
more accurate because jax2tf will both lower to StableHLO and then
serialize to StableHLO with compatibility guarantees.

This is part of launching the new version of jax2tf with native
serialization.

For now we keep also the parameter `experimental_native_lowering` and
the flag `jax2tf_default_experimental_native_lowering`, until we transition
projects using these flags to the new ones (separate change).

PiperOrigin-RevId: 516864636