Relax type of the `values` argument to .at[...].set(...) and friends.
The `values` gets passed to jnp.asarray(...) and there are users that are currently passing not just ArrayLikes but also things that can be coerced to arrays (e.g., tuples). Relax the type annotation to better describe the world as it is.

We may wish to enforce that the arguments are arrays in the future, but we should add a dynamic check/warning in that case.

PiperOrigin-RevId: 517451811